Semantically invert repeat button
It now works as a toggle, where the current icon is that of the current
state, rather than the state that clicking the button would bring the
application into.

Additionally, upgrade to using DaisyUI's switch, for some subtle
transition that is more in line with those of the new coutdown card.